Splited cli file into commands
