+++ .gitignore
@@ -1,0 +1,6 @@
+.idea/.name
+.idea/Generamba.iml
+.idea/misc.xml
+.idea/modules.xml
+.idea/vcs.xml
+.idea/workspace.xml
